# backend/app/routes/admin/users.py
import asyncio
from datetime import datetime
from typing import Annotated, Any, cast
from uuid import UUID

from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Path, Query, Request
from postgrest.types import CountMethod

from app.dependencies import (
    get_async_supabase_admin_client,
    get_email_service,
)
from app.limiter import limiter
from app.logger import logger, sanitize_log_value
from app.middleware.auth_middleware import (
    invalidate_user_auth_cache,
    require_permission,
)
from app.routes.admin.helpers import ADMIN_ERROR_RESPONSES, CommonPagination
from app.schemas.admin_schemas import BulkUserAction, RoleUpdateAdmin, UserBan, UserUpdateAdmin
from app.schemas.user import User
from app.services.email_service import EmailService
from app.services.redis_service import redis_service
from app.services.subscription_service import cancel_customer_subscriptions
from app.services.token_service import get_token_service
from app.utils.audit_logger import log_admin_action
from app.utils.avatar import sanitize_avatar_url
from app.utils.security_alerts import track_bulk_operation

router = APIRouter()

UserIdPath = Annotated[UUID, Path(title="The ID of the user", description="Must be a valid UUID")]


async def _invalidate_user_list_cache() -> None:
    """Invalidate all admin user list cache pages after any mutation."""
    await redis_service.delete_pattern("admin_users:*")


def _sanitize_user_picture(user_data: dict[str, Any]) -> dict[str, Any]:
    """Prevent legacy or manually altered avatar data from reaching clients."""
    sanitized = user_data.copy()
    sanitized["picture"] = sanitize_avatar_url(sanitized.get("picture"))
    return sanitized


ADMIN_USER_SORT_FIELDS = {"created_at", "email", "name", "treat_balance", "role"}
USER_NOT_FOUND_DETAIL = "User not found"


def _build_user_list_query(admin_client: Any, pagination: CommonPagination, search: str | None) -> Any:
    query = (
        admin_client.table("users")
        .select(
            "id, email, name, picture, treat_balance, is_pro, created_at, banned_at, roles(name)",
            count=CountMethod.exact,
        )
        .range(pagination.offset, pagination.offset + pagination.limit - 1)
    )
    sort_field = pagination.sort_by if pagination.sort_by in ADMIN_USER_SORT_FIELDS else "created_at"
    if sort_field != "role":
        query = query.order(sort_field, desc=pagination.order == "desc")
    if search:
        clean_search = "".join(c for c in search if c.isalnum() or c in (" ", "_", "-", "@", ".")).strip()
        if clean_search:
            query = cast(Any, query.or_(f"email.ilike.%{clean_search}%,name.ilike.%{clean_search}%"))
    return query


def _normalize_admin_user(user: dict[str, Any]) -> dict[str, Any]:
    user_copy = user.copy()
    role_info = user_copy.pop("roles", None)
    if isinstance(role_info, dict):
        role_name = role_info.get("name", "user")
    elif isinstance(role_info, list) and role_info:
        role_name = role_info[0].get("name", "user")
    else:
        role_name = "user"
    user_copy["role"] = role_name
    return _sanitize_user_picture(user_copy)


async def _invalidate_auth_cache_for_users(user_ids: list[str]) -> None:
    unique_user_ids = list(dict.fromkeys(user_ids))
    await asyncio.gather(*(invalidate_user_auth_cache(user_id) for user_id in unique_user_ids))


async def _invalidate_banned_user_auth_state(user_ids: list[str], reason: str) -> None:
    unique_user_ids = list(dict.fromkeys(user_ids))
    token_service = await get_token_service()

    async def invalidate_one(user_id: str) -> None:
        await asyncio.gather(
            invalidate_user_auth_cache(user_id),
            token_service.blacklist_all_user_tokens(user_id, reason=reason),
        )

    await asyncio.gather(*(invalidate_one(user_id) for user_id in unique_user_ids))


@router.get("/users", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("60/minute")
async def list_users(
    request: Request,
    pagination: Annotated[CommonPagination, Depends()],
    search: Annotated[str | None, Query()] = None,
    current_admin: Annotated[User | None, Depends(require_permission("users:read"))] = None,
) -> dict[str, Any]:
    """
    List all users with pagination, search, and sorting.
    """
    # Skip cache for search queries — results must be accurate and fresh.
    # Cache paginated list views for 60s to reduce DB load on the common case.
    cache_key = (
        f"admin_users:{pagination.offset}:{pagination.limit}:{pagination.sort_by}:{pagination.order}"
        if not search
        else None
    )
    if cache_key:
        cached = await redis_service.get(cache_key)
        if cached:
            cached_result = cast(dict[str, Any], cached)
            cached_result["data"] = [
                _sanitize_user_picture(user) for user in cast(list[dict[str, Any]], cached_result.get("data", []))
            ]
            return cached_result
    try:
        admin_client = await get_async_supabase_admin_client()
        result = await _build_user_list_query(admin_client, pagination, search).execute()
        users_data = cast(list[dict[str, Any]], result.data or [])
        total_count = result.count if result.count is not None else 0
        processed_data = [_normalize_admin_user(user) for user in users_data]

        result_data = {"data": processed_data, "total": total_count}
        if cache_key:
            await redis_service.set(cache_key, result_data, expire=60)  # 60-second TTL
        return result_data
    except Exception as e:
        logger.error("Failed to list users: %s", e)
        raise HTTPException(status_code=500, detail="Failed to fetch users")


@router.post("/users/bulk-ban", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("5/minute")
async def bulk_ban_users(
    request: Request,
    action: BulkUserAction,
    background_tasks: BackgroundTasks,
    current_admin: Annotated[User, Depends(require_permission("users:update"))],
    email_service: Annotated[EmailService, Depends(get_email_service)],
) -> dict[str, Any]:
    """
    Ban multiple users at once.
    """
    try:
        admin_client = await get_async_supabase_admin_client()
        user_ids_str = [str(uid) for uid in action.user_ids]

        # Robust check to protect admins
        roles_check = (
            await admin_client.table("users").select("id, email, roles(name)").in_("id", user_ids_str).execute()
        )

        target_uids = []
        target_emails = []
        skipped_admins = 0

        role_rows = cast(list[dict[str, Any]], roles_check.data or [])
        for u in role_rows:
            role_name = (u.get("roles") or {}).get("name") or "user"
            if role_name.lower() in ("admin", "super_admin"):
                skipped_admins += 1
            else:
                target_uids.append(u["id"])
                if u.get("email"):
                    target_emails.append(u["email"])

        if not target_uids:
            return {"message": "No valid users found to ban.", "skipped": skipped_admins}

        # SECURITY: Track bulk operations for alerting
        ip_address = request.client.host if request.client else "unknown"
        track_bulk_operation(
            user_id=current_admin.id,
            operation="bulk_ban",
            record_count=len(target_uids),
            ip_address=ip_address,
        )

        # Perform bulk update
        await (
            admin_client.table("users")
            .update({"banned_at": datetime.now().isoformat()})
            .in_("id", target_uids)
            .execute()
        )
        await _invalidate_banned_user_auth_state(target_uids, reason="bulk_admin_ban")

        # Background task for emails
        for email in target_emails:
            background_tasks.add_task(
                email_service.send_ban_notification, to_email=email, reason=action.reason or "Bulk Administrative Ban"
            )

        # Log Audit
        await log_admin_action(
            admin_client=admin_client,
            admin_id=current_admin.id,
            action="BULK_BAN",
            target_type="users",
            target_id="multiple",
            details={"target_user_ids": target_uids, "reason": action.reason, "ip": ip_address},
        )

        await _invalidate_user_list_cache()
        return {"message": f"Successfully banned {len(target_uids)} users.", "skipped": skipped_admins}
    except Exception as e:
        logger.error("Bulk ban failed: %s", e)
        raise HTTPException(status_code=500, detail="Bulk action failed")


@router.post("/users/bulk-unban", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("5/minute")
async def bulk_unban_users(
    request: Request,
    action: BulkUserAction,
    current_admin: Annotated[User, Depends(require_permission("users:update"))],
) -> dict[str, Any]:
    """
    Unban multiple users at once.
    """
    try:
        admin_client = await get_async_supabase_admin_client()
        user_ids_str = [str(uid) for uid in action.user_ids]

        await admin_client.table("users").update({"banned_at": None}).in_("id", user_ids_str).execute()
        await _invalidate_auth_cache_for_users(user_ids_str)

        # Log Audit
        await log_admin_action(
            admin_client=admin_client,
            admin_id=current_admin.id,
            action="BULK_UNBAN",
            target_type="users",
            target_id="multiple",
            details={"target_user_ids": user_ids_str, "ip": request.client.host if request.client else "unknown"},
        )

        await _invalidate_user_list_cache()
        return {"message": f"Successfully unbanned {len(user_ids_str)} users."}
    except Exception as e:
        logger.error("Bulk unban failed: %s", e)
        raise HTTPException(status_code=500, detail="Bulk action failed")


@router.delete("/users/{user_id}", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("10/minute")
async def delete_user(
    request: Request,
    user_id: UserIdPath,
    background_tasks: BackgroundTasks,
    current_admin: Annotated[User, Depends(require_permission("users:delete"))],
    email_service: Annotated[EmailService, Depends(get_email_service)],
) -> dict[str, str]:
    """
    Permanently delete a user and their data.
    """
    user_id_str = str(user_id)
    try:
        admin_client = await get_async_supabase_admin_client()

        check = (
            await admin_client.table("users")
            .select("email, stripe_customer_id, roles(name)")
            .eq("id", user_id_str)
            .execute()
        )
        if not check.data:
            raise HTTPException(status_code=404, detail=USER_NOT_FOUND_DETAIL)

        user_data = cast(dict[str, Any], check.data[0])
        role_info = user_data.get("roles")
        role_name = (role_info.get("name") if role_info else "user").lower()

        if role_name == "admin" or role_name == "super_admin":
            raise HTTPException(status_code=400, detail="Cannot delete an admin user")

        customer_id = user_data.get("stripe_customer_id")
        if isinstance(customer_id, str) and customer_id:
            await cancel_customer_subscriptions(customer_id)

        if user_data.get("email"):
            background_tasks.add_task(
                email_service.send_account_deletion_notification,
                to_email=user_data["email"],
                reason="Administrative Action",
            )

        anonymized_data = {
            "name": "[Deleted User]",
            "email": f"deleted_{user_id_str}@example.com",
            "picture": None,
            "bio": None,
            "stripe_customer_id": None,
        }

        await admin_client.table("users").update(anonymized_data).eq("id", user_id_str).execute()
        await _invalidate_banned_user_auth_state([user_id_str], reason="admin_delete")

        # Log action
        await log_admin_action(
            admin_client=admin_client,
            admin_id=current_admin.id,
            action="SOFT_DELETE_USER",
            target_type="users",
            target_id=user_id_str,
            details={
                "original_email": user_data.get("email"),
                "ip": request.client.host if request.client else "unknown",
            },
        )

        await _invalidate_user_list_cache()
        return {"message": f"User {user_id_str} data anonymized"}
    except HTTPException:
        raise
    except Exception:
        logger.error("Failed to delete user %s", sanitize_log_value(user_id_str))
        raise HTTPException(status_code=500, detail="Failed to delete user")


@router.patch("/users/{user_id}/profile", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("20/minute")
async def update_user_profile_admin(
    request: Request,
    user_id: UserIdPath,
    update_data: UserUpdateAdmin,
    current_admin: Annotated[User, Depends(require_permission("users:write"))],
) -> dict[str, Any]:
    """
    Update user profile as admin.
    """
    user_id_str = str(user_id)
    try:
        admin_client = await get_async_supabase_admin_client()
        filtered_data = update_data.model_dump(exclude_unset=True)

        if not filtered_data:
            raise HTTPException(status_code=400, detail="No valid fields provided")

        if "picture" in filtered_data:
            from app.utils.avatar import validate_avatar_url

            try:
                filtered_data["picture"] = validate_avatar_url(filtered_data["picture"])
            except ValueError as exc:
                raise HTTPException(status_code=400, detail=str(exc)) from exc

        result = await admin_client.table("users").update(filtered_data).eq("id", user_id_str).execute()
        if not result.data:
            raise HTTPException(status_code=404, detail=USER_NOT_FOUND_DETAIL)
        await invalidate_user_auth_cache(user_id_str)
        await _invalidate_user_list_cache()

        return _sanitize_user_picture(cast(dict[str, Any], result.data[0]))
    except HTTPException:
        raise
    except Exception as e:
        logger.error("Failed to update user profile %s: %s", sanitize_log_value(user_id_str), e)
        raise HTTPException(status_code=500, detail="Failed to update user profile")


@router.put("/users/{user_id}/role", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("20/minute")
async def update_user_role(
    request: Request,
    user_id: UserIdPath,
    role_data: RoleUpdateAdmin,
    current_admin: Annotated[User, Depends(require_permission("users:update"))],
) -> dict[str, Any]:
    """
    Update user role.
    """
    user_id_str = str(user_id)
    role_id_str = str(role_data.role_id) if role_data.role_id else None
    if not role_id_str:
        raise HTTPException(status_code=400, detail="role_id is required")

    try:
        admin_client = await get_async_supabase_admin_client()
        role_check = await admin_client.table("roles").select("name").eq("id", role_id_str).single().execute()
        if not role_check.data:
            raise HTTPException(status_code=404, detail="Role not found")

        role_record = cast(dict[str, Any], role_check.data)
        role_name = cast(str | None, role_record.get("name"))
        result = await admin_client.table("users").update({"role_id": role_id_str}).eq("id", user_id_str).execute()
        if not result.data:
            raise HTTPException(status_code=404, detail=USER_NOT_FOUND_DETAIL)
        await invalidate_user_auth_cache(user_id_str)
        await _invalidate_user_list_cache()

        # Log Audit
        await log_admin_action(
            admin_client=admin_client,
            admin_id=current_admin.id,
            action="UPDATE_ROLE",
            target_type="users",
            target_id=user_id_str,
            details={"new_role": role_name, "ip": request.client.host if request.client else "unknown"},
        )

        return cast(dict[str, Any], result.data[0])
    except HTTPException:
        raise
    except Exception:
        logger.error("Failed to update user role %s", sanitize_log_value(user_id_str))
        raise HTTPException(status_code=500, detail="Failed to update role")


@router.post("/users/{user_id}/ban", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("5/minute")
async def ban_user(
    request: Request,
    user_id: UserIdPath,
    ban_data: UserBan,
    background_tasks: BackgroundTasks,
    current_admin: Annotated[User, Depends(require_permission("users:update"))],
    email_service: Annotated[EmailService, Depends(get_email_service)],
) -> dict[str, str]:
    """
    Ban a user.
    """
    user_id_str = str(user_id)
    try:
        admin_client = await get_async_supabase_admin_client()
        supa_res = (
            await admin_client.table("users")
            .select("id, email, name, username, picture, bio, created_at, banned_at, roles(name)")
            .eq("id", user_id)
            .single()
            .execute()
        )
        user_data = cast(dict[str, Any], supa_res.data)
        if not user_data:
            raise HTTPException(status_code=404, detail=USER_NOT_FOUND_DETAIL)

        role_name = (user_data.get("roles") or {}).get("name") or "user"
        if role_name.lower() in ("admin", "super_admin"):
            raise HTTPException(status_code=400, detail="Cannot ban an admin user")

        await (
            admin_client.table("users")
            .update({"banned_at": datetime.now().isoformat()})
            .eq("id", user_id_str)
            .execute()
        )
        await _invalidate_banned_user_auth_state([user_id_str], reason="admin_ban")

        if user_data.get("email"):
            background_tasks.add_task(
                email_service.send_ban_notification, to_email=user_data["email"], reason=ban_data.reason
            )

        # Log Audit
        await log_admin_action(
            admin_client=admin_client,
            admin_id=current_admin.id,
            action="BAN_USER",
            target_type="users",
            target_id=user_id_str,
            details={"reason": ban_data.reason, "ip": request.client.host if request.client else "unknown"},
        )

        await _invalidate_user_list_cache()
        return {"message": "User banned"}
    except HTTPException:
        raise
    except Exception:
        logger.error("Failed to ban user")
        raise HTTPException(status_code=500, detail="Failed to ban user")


@router.post("/users/{user_id}/unban", responses=ADMIN_ERROR_RESPONSES)
@limiter.limit("5/minute")
async def unban_user(
    request: Request,
    user_id: UserIdPath,
    current_admin: Annotated[User, Depends(require_permission("users:update"))],
) -> dict[str, str]:
    """
    Unban a user.
    """
    user_id_str = str(user_id)
    try:
        admin_client = await get_async_supabase_admin_client()
        await admin_client.table("users").update({"banned_at": None}).eq("id", user_id_str).execute()
        await invalidate_user_auth_cache(user_id_str)

        # Log Audit
        await log_admin_action(
            admin_client=admin_client,
            admin_id=current_admin.id,
            action="UNBAN_USER",
            target_type="users",
            target_id=user_id_str,
            details={"ip": request.client.host if request.client else "unknown"},
        )

        await _invalidate_user_list_cache()
        return {"message": "User unbanned"}
    except Exception:
        logger.error("Failed to unban user")
        raise HTTPException(status_code=500, detail="Failed to unban user")