# docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PurrFect Spots API",
    "description": "\n        PurrFect Spots API helps you share and discover cat locations in a socially-connected way.\n\n        ### 🌐 API Specification\n        All endpoints follow the `/api/v1/` prefix convention for stable identification.\n\n        ### ✨ Key Features\n        * 📍 **Geo-Location Awareness**: High-precision cat spot sharing with privacy protection.\n        * 🤳 **AI Photo Intelligence**: Automated cat detection and visual attribute extraction.\n        * 🔒 **Enterprise-Grade Security**: PKCE-enabled OAuth 2.0 and Supabase-backed authentication.\n        * ⚡ **Performance Optimized**: Built-in rate limiting, GZip compression, and ETag support.\n        * 🔄 **Real-time Ready**: Designed for low-latency interactions and live updates.\n        ",
    "contact": {
      "name": "Purrfect Spots Team",
      "email": "support@purrfectspots.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "3.1.0"
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Users",
        "description": "List all users with pagination, search, and sorting.",
        "operationId": "list_users_api_v1_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/bulk-ban": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Bulk Ban Users",
        "description": "Ban multiple users at once.",
        "operationId": "bulk_ban_users_api_v1_admin_users_bulk_ban_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUserAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Bulk Ban Users Api V1 Admin Users Bulk Ban Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/bulk-unban": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Bulk Unban Users",
        "description": "Unban multiple users at once.",
        "operationId": "bulk_unban_users_api_v1_admin_users_bulk_unban_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUserAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Bulk Unban Users Api V1 Admin Users Bulk Unban Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete User",
        "description": "Permanently delete a user and their data.",
        "operationId": "delete_user_api_v1_admin_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "The ID of the user",
              "description": "Must be a valid UUID"
            },
            "description": "Must be a valid UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete User Api V1 Admin Users  User Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/profile": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update User Profile Admin",
        "description": "Update user profile as admin.",
        "operationId": "update_user_profile_admin_api_v1_admin_users__user_id__profile_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "The ID of the user",
              "description": "Must be a valid UUID"
            },
            "description": "Must be a valid UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateAdmin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update User Profile Admin Api V1 Admin Users  User Id  Profile Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/role": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update User Role",
        "description": "Update user role.",
        "operationId": "update_user_role_api_v1_admin_users__user_id__role_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "The ID of the user",
              "description": "Must be a valid UUID"
            },
            "description": "Must be a valid UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateAdmin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update User Role Api V1 Admin Users  User Id  Role Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/ban": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Ban User",
        "description": "Ban a user.",
        "operationId": "ban_user_api_v1_admin_users__user_id__ban_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "The ID of the user",
              "description": "Must be a valid UUID"
            },
            "description": "Must be a valid UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserBan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ban User Api V1 Admin Users  User Id  Ban Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/unban": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Unban User",
        "description": "Unban a user.",
        "operationId": "unban_user_api_v1_admin_users__user_id__unban_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "The ID of the user",
              "description": "Must be a valid UUID"
            },
            "description": "Must be a valid UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Unban User Api V1 Admin Users  User Id  Unban Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/roles": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Roles",
        "description": "List all available roles.",
        "operationId": "list_roles_api_v1_admin_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Roles Api V1 Admin Roles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleUpdateAdmin": {
        "properties": {
          "role_id": {
            "type": "string",
            "title": "Role Id",
            "description": "UUID of the role"
          }
        },
        "type": "object",
        "required": [
          "role_id"
        ],
        "title": "RoleUpdateAdmin"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}